refactor: ExceptionCode -> ErrorCode 롤백
