+++ backend/src/main/java/com/woowacourse/f12/dto/response/ExceptionResponse.java
@@ -1,28 +1,28 @@
 package com.woowacourse.f12.dto.response;
 
 import com.woowacourse.f12.exception.CustomException;
-import com.woowacourse.f12.exception.ExceptionCode;
+import com.woowacourse.f12.exception.ErrorCode;
 import lombok.Getter;
 
 @Getter
 public class ExceptionResponse {
 
     private String message;
-    private ExceptionCode exceptionCode;
+    private ErrorCode errorCode;
 
     private ExceptionResponse() {
     }
 
-    private ExceptionResponse(final String message, final ExceptionCode exceptionCode) {
+    private ExceptionResponse(final String message, final ErrorCode errorCode) {
         this.message = message;
-        this.exceptionCode = exceptionCode;
+        this.errorCode = errorCode;
     }
 
     public static ExceptionResponse from(final CustomException e) {
-        return new ExceptionResponse(e.getMessage(), e.getExceptionCode());
+        return new ExceptionResponse(e.getMessage(), e.getErrorCode());
     }
 
-    public static ExceptionResponse from(final String message, final ExceptionCode exceptionCode) {
-        return new ExceptionResponse(message, exceptionCode);
+    public static ExceptionResponse from(final String message, final ErrorCode errorCode) {
+        return new ExceptionResponse(message, errorCode);
     }
 }
