+++ backend/src/main/java/com/woowacourse/f12/exception/badrequest/AlreadyWrittenReviewException.java
@@ -1,6 +1,6 @@
 package com.woowacourse.f12.exception.badrequest;
 
-import static com.woowacourse.f12.exception.ExceptionCode.DUPLICATED_DATA;
+import static com.woowacourse.f12.exception.ErrorCode.DUPLICATED_DATA;
 
 public class AlreadyWrittenReviewException extends InvalidValueException {
