+++ backend/src/main/java/com/woowacourse/f12/exception/badrequest/DuplicatedProfileProductCategoryException.java
@@ -1,6 +1,6 @@
 package com.woowacourse.f12.exception.badrequest;
 
-import static com.woowacourse.f12.exception.ExceptionCode.INVALID_REQUEST_BODY;
+import static com.woowacourse.f12.exception.ErrorCode.INVALID_REQUEST_BODY;
 
 public class DuplicatedProfileProductCategoryException extends InvalidValueException {
