+++ backend/src/main/java/com/woowacourse/f12/exception/badrequest/InvalidJobTypeException.java
@@ -1,10 +1,10 @@
 package com.woowacourse.f12.exception.badrequest;
 
-import com.woowacourse.f12.exception.ExceptionCode;
+import com.woowacourse.f12.exception.ErrorCode;
 
 public class InvalidJobTypeException extends InvalidValueException {
 
     public InvalidJobTypeException() {
-        super(ExceptionCode.INVALID_SEARCH_PARAM, "올바르지 않은 직군 입력입니다.");
+        super(ErrorCode.INVALID_SEARCH_PARAM, "올바르지 않은 직군 입력입니다.");
     }
 }
