+++ backend/src/main/java/com/woowacourse/f12/exception/badrequest/InvalidValueException.java
@@ -1,11 +1,11 @@
 package com.woowacourse.f12.exception.badrequest;
 
 import com.woowacourse.f12.exception.CustomException;
-import com.woowacourse.f12.exception.ExceptionCode;
+import com.woowacourse.f12.exception.ErrorCode;
 
 public class InvalidValueException extends CustomException {
 
-    public InvalidValueException(final ExceptionCode exceptionCode, final String message) {
-        super(exceptionCode, message);
+    public InvalidValueException(final ErrorCode errorCode, final String message) {
+        super(errorCode, message);
     }
 }
