+++ backend/src/main/java/com/woowacourse/f12/exception/forbidden/ForbiddenMemberException.java
@@ -1,11 +1,11 @@
 package com.woowacourse.f12.exception.forbidden;
 
 import com.woowacourse.f12.exception.CustomException;
-import com.woowacourse.f12.exception.ExceptionCode;
+import com.woowacourse.f12.exception.ErrorCode;
 
 public class ForbiddenMemberException extends CustomException {
 
-    public ForbiddenMemberException(final ExceptionCode exceptionCode, final String message) {
-        super(exceptionCode, message);
+    public ForbiddenMemberException(final ErrorCode errorCode, final String message) {
+        super(errorCode, message);
     }
 }
