+++ backend/src/main/java/com/woowacourse/f12/exception/internalserver/ExternalServerException.java
@@ -1,11 +1,11 @@
 package com.woowacourse.f12.exception.internalserver;
 
 import com.woowacourse.f12.exception.CustomException;
-import com.woowacourse.f12.exception.ExceptionCode;
+import com.woowacourse.f12.exception.ErrorCode;
 
 public class ExternalServerException extends CustomException {
 
-    public ExternalServerException(final ExceptionCode exceptionCode, final String message) {
-        super(exceptionCode, message);
+    public ExternalServerException(final ErrorCode errorCode, final String message) {
+        super(errorCode, message);
     }
 }
