+++ backend/src/main/java/com/woowacourse/f12/exception/internalserver/GitHubServerException.java
@@ -1,6 +1,6 @@
 package com.woowacourse.f12.exception.internalserver;
 
-import static com.woowacourse.f12.exception.ExceptionCode.EXTERNAL_SERVER_ERROR;
+import static com.woowacourse.f12.exception.ErrorCode.EXTERNAL_SERVER_ERROR;
 
 public class GitHubServerException extends ExternalServerException {
