+++ backend/src/main/java/com/woowacourse/f12/exception/notfound/NotFoundException.java
@@ -1,11 +1,11 @@
 package com.woowacourse.f12.exception.notfound;
 
 import com.woowacourse.f12.exception.CustomException;
-import com.woowacourse.f12.exception.ExceptionCode;
+import com.woowacourse.f12.exception.ErrorCode;
 
 public class NotFoundException extends CustomException {
 
-    public NotFoundException(final ExceptionCode exceptionCode, final String message) {
-        super(exceptionCode, message);
+    public NotFoundException(final ErrorCode errorCode, final String message) {
+        super(errorCode, message);
     }
 }
