+++ backend/src/main/java/com/woowacourse/f12/exception/unauthorized/TokenExpiredException.java
@@ -1,6 +1,6 @@
 package com.woowacourse.f12.exception.unauthorized;
 
-import static com.woowacourse.f12.exception.ExceptionCode.TOKEN_EXPIRED;
+import static com.woowacourse.f12.exception.ErrorCode.TOKEN_EXPIRED;
 
 public class TokenExpiredException extends UnauthorizedException {
